move StorageUtils functions to Slot data structure
